refactor(gateway): GAR-691 — Q10.g extract build_telegram_channels to bootstrap/telegram.rs
Extracts build_telegram_voice_handler + build_telegram_channels + handle_command
(~423 LOC, Telegram wiring + allowlist + pairing + voice handler) from
bootstrap/mod.rs into a new sibling module bootstrap/telegram.rs.
Re-exports build_telegram_channels at the same public path — zero behaviour change.
bootstrap/mod.rs: 1579 → 1146 LOC (−433). Removes telegram-specific imports
(Mutex, OnVoiceFn, TelegramChannel, Allowlist, PairingManager, teloxide, ChatMessage,
error!) from mod.rs. Adds plans/0170 and plans/README.md row.
