+++ plans/README.md
@@ -178,4 +178,5 @@ Este diretório contém os planos de implementação para o projeto GarraRUST.
 | 0166 | [GAR-478 — Q10.d: extract `bootstrap/slack.rs` (166 LOC, zero behaviour change)](0166-gar-478-q10d-bootstrap-slack.md) | [GAR-478](https://linear.app/chatgpt25/issue/GAR-478) | ✅ Merged 2026-05-22 via PR #474 (`4a51841`) |
 | 0167 | [GAR-479 — Q10.e: extract `bootstrap/whatsapp.rs` (180 LOC, zero behaviour change)](0167-gar-479-q10e-bootstrap-whatsapp.md) | [GAR-479](https://linear.app/chatgpt25/issue/GAR-479) | ✅ Merged 2026-05-22 via PR #476 (`60a8dff`) |
 | 0168 | [GAR-480 — Q10.f: extract `bootstrap/imessage.rs` (123 LOC, macOS-only, zero behaviour change)](0168-gar-480-q10f-bootstrap-imessage.md) | [GAR-480](https://linear.app/chatgpt25/issue/GAR-480) | 🔄 In Progress |
-| 0169 | [GAR-513 — deny.toml advisory-not-detected cleanup: remove stale RUSTSEC-2024-0429 + RUSTSEC-2026-0097 from deny.toml, retain in audit.toml](0169-gar-513-deny-toml-advisory-not-detected.md) | [GAR-513](https://linear.app/chatgpt25/issue/GAR-513) | 🚧 In Progress |
+| 0169 | [GAR-513 — deny.toml advisory-not-detected cleanup: remove stale RUSTSEC-2024-0429 + RUSTSEC-2026-0097 from deny.toml, retain in audit.toml](0169-gar-513-deny-toml-advisory-not-detected.md) | [GAR-513](https://linear.app/chatgpt25/issue/GAR-513) | ✅ Merged 2026-05-23 via PR #484 (`b3f62fd`) |
+| 0170 | [GAR-691 — Q10.g: extract `bootstrap/telegram.rs` (423 LOC, zero behaviour change)](0170-gar-691-q10g-bootstrap-telegram.md) | [GAR-691](https://linear.app/chatgpt25/issue/GAR-691) | 🔄 In Progress |
